move observation e2e logic to a page object
